+++ eslint/src/rules/4/deprecated.alert-action-params.ts
@@ -1,0 +1,9 @@
+import {buildRule} from '../../core/deprecated-alias';
+
+/**
+ * Rule for deprecated `AlertActionParams` alias for {@link ESLAlertActionParams}
+ */
+export default buildRule({
+  alias: 'ESLAlertActionParams',
+  deprecation: 'AlertActionParams'
+});