+++ eslint/src/rules/4/deprecated.popup-action-params.ts
@@ -1,0 +1,9 @@
+import {buildRule} from '../../core/deprecated-alias';
+
+/**
+ * Rule for deprecated `PopupActionParams` alias for {@link ESLPopupActionParams}
+ */
+export default buildRule({
+  alias: 'ESLPopupActionParams',
+  deprecation: 'PopupActionParams'
+});