+++ eslint/src/rules/4/deprecated.toggleable-action-params.ts
@@ -5,5 +5,5 @@ import {buildRule} from '../../core/deprecated-alias';
  */
 export default buildRule({
   alias: 'ESLToggleableActionParams',
-  deprecation: 'ToggleableActionParams',
+  deprecation: 'ToggleableActionParams'
 });