+++ src/modules/all.ts
@@ -24,6 +24,9 @@ export * from './esl-panel-group/core';
 
 export * from './esl-popup/core';
 
+// Open State Mixin
+export * from './esl-open-state/core';
+
 // Scrollbar
 export * from './esl-scrollbar/core';