+++ src/modules/esl-alert/test/alert.request.test.ts
@@ -1,6 +1,6 @@
 import {ESLEventUtils} from '../../esl-event-listener/core/api';
 import {ESLAlert} from '../core/esl-alert';
-import type {AlertActionParams} from '../core/esl-alert';
+import type {ESLAlertActionParams} from '../core/esl-alert';
 
 describe('ESLAlert: general', () => {
   beforeAll(() => ESLAlert.register());
@@ -32,7 +32,7 @@ describe('ESLAlert: events', () => {
   });
 
   test('Show/Hide events ', () => {
-    ESLEventUtils.dispatch($el, ESLAlert.prototype.SHOW_REQUEST_EVENT);
+    ESLEventUtils.dispatch($el, ESLAlert.prototype.SHOW_REQUEST_EVENT, {detail: {text: 'esl-alert-text'}});
     jest.advanceTimersByTime(1);
     expect($el.open).toBe(true);
     ESLEventUtils.dispatch($el, ESLAlert.prototype.HIDE_REQUEST_EVENT);
@@ -41,7 +41,7 @@ describe('ESLAlert: events', () => {
   });
 
   test('Auto hide by hideTime & hideDelay ', () => {
-    const params: CustomEventInit<AlertActionParams> = {detail: {hideTime: 100, hideDelay: 100}};
+    const params: CustomEventInit<ESLAlertActionParams> = {detail: {hideTime: 100, hideDelay: 100, text: 'esl-alert-text'}};
     ESLEventUtils.dispatch($el, ESLAlert.prototype.SHOW_REQUEST_EVENT, params);
     jest.advanceTimersByTime(1);
     expect($el.open).toBe(true);
@@ -50,19 +50,19 @@ describe('ESLAlert: events', () => {
   });
 
   test('Add text params', () => {
-    const params: CustomEventInit<AlertActionParams> = {detail: {text: 'esl-alert-text'}};
+    const params: CustomEventInit<ESLAlertActionParams> = {detail: {text: 'esl-alert-text'}};
     ESLEventUtils.dispatch($el, ESLAlert.prototype.SHOW_REQUEST_EVENT, params);
     expect('esl-alert-text').toEqual($el.textContent);
   });
 
   test('Add HTML params', () => {
-    const params: CustomEventInit<AlertActionParams> = {detail: {html: '<h2>esl-alert-html</h2>'}};
+    const params: CustomEventInit<ESLAlertActionParams> = {detail: {html: '<h2>esl-alert-html</h2>'}};
     ESLEventUtils.dispatch($el, ESLAlert.prototype.SHOW_REQUEST_EVENT, params);
     expect('<h2>esl-alert-html</h2>').toEqual($el.children[0].innerHTML);
   });
 
   test('Add class params', () => {
-    const params: CustomEventInit<AlertActionParams> = {detail: {cls: 'esl-alert-cls',  text: 'esl-alert-text'}};
+    const params: CustomEventInit<ESLAlertActionParams> = {detail: {cls: 'esl-alert-cls',  text: 'esl-alert-text'}};
     ESLEventUtils.dispatch($el, ESLAlert.prototype.SHOW_REQUEST_EVENT, params);
     expect($el.className).toContain('esl-alert-cls');
   });