+++ src/modules/esl-toggleable/core/esl-toggleable.ts
@@ -14,6 +14,8 @@ import type {DelegatedEvent} from '../../esl-event-listener/core/types';
 
 /** Default Toggleable action params type definition */
 export interface ESLToggleableActionParams {
+  /** Action to execute */
+  readonly action?: 'show' | 'hide';
   /** Initiator string identifier */
   initiator?: string;
   /** Delay timeout for both show and hide actions */
@@ -130,6 +132,13 @@ export class ESLToggleable extends ESLBaseElement {
   /** Inner state */
   private _open: boolean = false;
 
+  /**
+   * Marker of last opened state.
+   * @deprecated and will be removed in 5.0.0 (`this.open` mutation will be moved inside `onShow` and `onHide` methods,
+   * so it became possible to track the initial state of the toggleable inside of action hooks)
+   */
+  protected wasOpened: boolean = false;
+
   /** Inner show/hide task manager instance */
   protected _task: DelayedTask = new DelayedTask();
   /** Marker for current hover listener state */
@@ -220,17 +229,21 @@ export class ESLToggleable extends ESLBaseElement {
 
   /** Actual show task to execute by toggleable task manger ({@link DelayedTask} out of the box) */
   protected showTask(params: ESLToggleableActionParams): void {
+    Object.defineProperty(params, 'action', {value: 'show', writable: false});
     if (!this.shouldShow(params)) return;
     if (!params.silent && !this.$$fire(this.BEFORE_SHOW_EVENT, {detail: {params}})) return;
     this.activator = params.activator;
+    this.wasOpened = this.open;
     this.open = true;
     this.onShow(params);
     if (!params.silent) this.$$fire(this.SHOW_EVENT, {detail: {params}, cancelable: false});
   }
   /** Actual hide task to execute by toggleable task manger ({@link DelayedTask} out of the box) */
   protected hideTask(params: ESLToggleableActionParams): void {
+    Object.defineProperty(params, 'action', {value: 'hide', writable: false});
     if (!this.shouldHide(params)) return;
     if (!params.silent && !this.$$fire(this.BEFORE_HIDE_EVENT, {detail: {params}})) return;
+    this.wasOpened = this.open;
     this.open = false;
     this.onHide(params);
     this.bindOutsideEventTracking(false);
@@ -242,8 +255,15 @@ export class ESLToggleable extends ESLBaseElement {
    * Returns false if the show action should not be executed.
    */
   protected shouldShow(params: ESLToggleableActionParams): boolean {
+    if (typeof this.onBeforeShow === 'function') return this.onBeforeShow(params) === false;
     return params.force || !this.open;
   }
+
+  /**
+   * Called before show action to check if the action should be executed.
+   * @deprecated use {@link shouldShow} and immutable aproach instead
+   */
+  protected onBeforeShow?: (params: ESLToggleableActionParams) => boolean | void;
 
   /**
    * Actions to execute on show toggleable.
@@ -267,8 +287,15 @@ export class ESLToggleable extends ESLBaseElement {
    * Returns false if the hide action should not be executed.
    */
   protected shouldHide(params: ESLToggleableActionParams): boolean {
+    if (typeof this.onBeforeHide === 'function') return this.onBeforeHide(params) === false;
     return params.force || this.open;
   }
+
+  /**
+   * Called before hide action to check if the action should be executed.
+   * @deprecated use {@link shouldHide} and immutable aproach instead
+   */
+  protected onBeforeHide?: (params: ESLToggleableActionParams) => boolean | void;
 
   /**
    * Actions to execute on hide toggleable.
@@ -338,7 +365,7 @@ export class ESLToggleable extends ESLBaseElement {
   @listen({
     auto: false,
     event: 'keydown mouseup touchend',
-    target: document.body,
+    target: document,
     capture: true
   })
   protected _onOutsideAction(e: Event): void {